fix an error, added formatting for BIOS BOOT partition

+++ arch-installer.sh
@@ -48,7 +48,7 @@ ls /sys/firmware/efi/efivars &> /dev/null
 if (( $? != 0 ))
 then
     echo -e "EFI system not detected, BIOS bootloader installation will be used \n"
-    echo -e "Grub will be used as bootloader"
+    echo -e "Grub will be used as bootloader \n"
     bootmode="BIOS"
     bootloader="grub"
     bootloader_extra=""
@@ -120,7 +120,7 @@ fatal_error $? "mounting the partition"
 
 while true
 do
-    echo -e "Path to EFI partition: \c"
+    echo -e "Path to BOOT partition: \c"
     read efi
 
     if [ ! -e $efi ]
@@ -135,11 +135,23 @@ do
     case $format_efi_selection in
         [Yy] )  echo
                 echo -e "The partition will be formatted and mounted \n"
-                mkfs.fat -F32 $efi
-                fatal_error $? "formatting the partition"
 
-                fatlabel $efi BOOT
-                fatal_error $? "assigning a label to the EFI partition"
+                if [[ bootmode == "EFI" ]]
+                then
+                    mkfs.fat -F32 $efi
+                    fatal_error $? "formatting the partition"
+
+                    fatlabel $efi BOOT
+                    fatal_error $? "assigning a label to the EFI partition"
+                
+                elif [[ bootmode == "BIOS" ]]
+                then
+                    wipefs -a $efi
+                    fatal_error $? "formatting the partition"
+
+                    parted $efi name N "BOOT"
+                    fatal_error $? "assignning a label to the BOOT partition"
+                fi
                 break;;
 
         [Nn] )  echo
@@ -270,7 +282,7 @@ arch-chroot /mnt /bin/bash -e << EOF
         then
             grub-install --recheck /dev/$(lsblk -ndo pkname $root)
         else
-            grub-install lsbl--target=x86_64-efi --efi-directory=/boot --bootloader-id=GRUB
+            grub-install --target=x86_64-efi --efi-directory=/boot --bootloader-id=GRUB
         fi
 
         sed -i "s|^GRUB_CMDLINE_LINUX_DEFAULT=\"loglevel=3 quiet\"|GRUB_CMDLINE_LINUX_DEFAULT=\"loglevel=3\"|" /etc/default/grub
